Refined shutdown state tracking
