+++ Core/Src/u_can.c
@@ -233,7 +233,7 @@ void can_inbox(can_msg_t *message) {
     case CANID_SHUTDOWN:
         shutdown_as_read_by_bms_t bms = { 0 };
         receive_shutdown_as_read_by_bms(message, &bms);
-        update_shutdown(bms.shutdown);
+        update_bms_shutdown(bms.shutdown);
 
         /* If shutdown is active, cancel the RTDS sound if it's active. */
         if(bms.shutdown == true) {
