+++ Core/Src/u_threads.c
@@ -21,6 +21,7 @@
 #include "u_lightning.h"
 #include "u_rtds.h"
 #include "u_nx_debug.h"
+#include "u_shutdown.h"
 #include "timer.h"
 #include "debounce.h"
 #include "u_traceout_app.h"
@@ -43,13 +44,6 @@
 #define PRIO_vRTDS             3
 #define PRIO_vTest             3
 #define PRIO_vPeripherals      3
-
-
-
-// callback for when either bms or imd indicates a fault - sets flag via arg
-static void _lightning_board_status_callback(void *arg) {
-    *(bool *)arg = true;
-}
 
 /* Test Thread */
 static thread_t test_thread = {
@@ -350,9 +344,7 @@ static thread_t shutdown_thread = {
         .auto_start = TX_AUTO_START,      /* Auto Start */
         .sleep      = 500,                /* Sleep (in ticks) */
         .function   = vShutdown           /* Thread Function */
-    };
-
-#define LIGHTNING_BOARD_DEBOUNCE 500
+};
 
 void vShutdown(ULONG thread_input) {
     /* Debounce Timer */
@@ -361,54 +353,8 @@ void vShutdown(ULONG thread_input) {
 
     while(1) {
 
-        bool bms_gpio = (HAL_GPIO_ReadPin(BMS_GPIO_GPIO_Port, BMS_GPIO_Pin) == GPIO_PIN_SET);
-        bool bots_gpio = (HAL_GPIO_ReadPin(BOTS_GPIO_GPIO_Port, BOTS_GPIO_Pin) == GPIO_PIN_SET);
-        bool spare_gpio = (HAL_GPIO_ReadPin(SPARE_GPIO_GPIO_Port, SPARE_GPIO_Pin) == GPIO_PIN_SET);
-        bool bspd_gpio = (HAL_GPIO_ReadPin(BSPD_GPIO_GPIO_Port, BSPD_GPIO_Pin) == GPIO_PIN_SET);
-        bool hv_c = (HAL_GPIO_ReadPin(HV_C_GPIO_GPIO_Port, HV_C_GPIO_Pin) == GPIO_PIN_SET);
-        bool hvd_gpio = (HAL_GPIO_ReadPin(HVD_GPIO_GPIO_Port, HVD_GPIO_Pin) == GPIO_PIN_SET);
-        bool imd_gpio = (HAL_GPIO_ReadPin(IMD_GPIO_GPIO_Port, IMD_GPIO_Pin) == GPIO_PIN_SET);
-        bool ckpt_gpio = (HAL_GPIO_ReadPin(CKPT_GPIO_GPIO_Port, CKPT_GPIO_Pin) == GPIO_PIN_SET);
-        bool inertia_sw_gpio = (HAL_GPIO_ReadPin(INERTIA_SW_GPIO_GPIO_Port, INERTIA_SW_GPIO_Pin) == GPIO_PIN_SET);
-        bool tsms_gpio = (HAL_GPIO_ReadPin(TSMS_GPIO_GPIO_Port, TSMS_GPIO_Pin) == GPIO_PIN_SET);
-
-
-        //lightning status with debounce
-        bool lightning_fault = bms_gpio || imd_gpio;
-
-        // Always call debounce so the cancel path runs when the fault clears.
-        // The callback sets lightning_is_red=true via arg; we track it separately
-        // so RED is sent every loop iteration (not just once per debounce cycle).
-        debounce(lightning_fault, &lightning_status_timer, LIGHTNING_BOARD_DEBOUNCE, &_lightning_board_status_callback, &lightning_is_red);
-        if (!lightning_fault) {
-            lightning_is_red = false;
-            send_lightning_board_status(LIGHT_GREEN);
-        } else if (lightning_is_red) {
-            send_lightning_board_status(LIGHT_RED);
-        }
-
-        /* Send Shutdown Pins CAN message. */
-        send_shutdown_pins(
-            bms_gpio,
-            bots_gpio,
-            spare_gpio,
-            bspd_gpio,
-            hv_c,
-            hvd_gpio,
-            imd_gpio,
-            ckpt_gpio,
-            inertia_sw_gpio,
-            tsms_gpio,
-            0
-        );
-
-        bool shutdown_active = bms_gpio || bots_gpio || spare_gpio || bspd_gpio
-                                || hv_c || hvd_gpio || imd_gpio || ckpt_gpio
-                                || inertia_sw_gpio || tsms_gpio;
-
-        if (shutdown_active || (is_shutdown_active() == true)) { // if tsms is still on when shutdown is active, trigger fault
-            queue_send(&faults, &(fault_t){SHUTDOWN_FAULT}, TX_NO_WAIT);
-        }
+        /* Process shutdown. */
+        shutdown_process();
 
         /* Sleep Thread for specified number of ticks. */
         tx_thread_sleep(shutdown_thread.sleep);
